Extract OwnerView to separate component file
